mycode_python/requirements.txt
ansible==2.9.9
asyncio==3.4.3
asyncssh==1.18.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
ciscoconfparse==1.5.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
dnspython==1.16.0
future==0.18.2
idna==2.9
Jinja2==2.11.2
junos-eznc==2.4.1
lxml==4.5.1
MarkupSafe==1.1.1
napalm==3.0.1
ncclient==0.6.7
netaddr==0.7.19
netdev==0.9.3
netmiko==3.1.1
ntc-templates==1.4.2
paramiko==2.7.1
passlib==1.7.2
pycparser==2.20
pyeapi==0.8.3
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.4
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
scp==0.13.2
six==1.15.0
textfsm==1.1.0
transitions==0.8.1
urllib3==1.25.9
yamlordereddictloader==0.4.0

